♻️ refactor(contract): delete dead DebouncedSearch/MetadataSource and relocate ABS types to server
